backend/requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-storage
google-cloud-aiplatform
firebase-admin
python-multipart
python-dotenv
